+++ ksdl/src/ksdl/system/KClock.kt
@@ -3,27 +3,27 @@ package ksdl.system
 import sdl2.*
 
 class KClock {
-    val frequency: Long = SDL_GetPerformanceFrequency()
-    var start: Long = SDL_GetPerformanceCounter()
+    val frequency: ULong = SDL_GetPerformanceFrequency()
+    var start: ULong = SDL_GetPerformanceCounter()
         private set
 
     fun reset() {
         start = SDL_GetPerformanceCounter()
     }
 
-    fun elapsedTicks(): Long = SDL_GetPerformanceCounter() - start
+    fun elapsedTicks(): ULong = SDL_GetPerformanceCounter() - start
 
-    fun elapsedMillis(): Long {
-        val elapsed = elapsedTicks() * 1000
-        return (elapsed.toDouble() / frequency).toLong()
+    fun elapsedMillis(): ULong {
+        val elapsed = elapsedTicks() * 1000u
+        return (elapsed / frequency)
     }
 
-    fun elapsedMicros(): Long {
-        val elapsed = elapsedTicks() * 1000000
-        return (elapsed.toDouble() / frequency).toLong()
+    fun elapsedMicros(): ULong {
+        val elapsed = elapsedTicks() * 1000000u
+        return (elapsed / frequency)
     }
 
-    fun elapsedSeconds(): Long {
-        return (elapsedTicks().toDouble() / frequency).toLong()
+    fun elapsedSeconds(): ULong {
+        return (elapsedTicks() / frequency)
     }
 }
